Move extension logic to relation_internal.cc
